refactor: make handlers more modular
